finish jwt functions for encode, decode and vefify, creating a jwt.js file in middleware folder
